Added custom debug methods, kinda.
- Debug Build step now adds the method definition to a
  function-local-variable '__DEBUG_METHOD'.
- Added __DEBUG_METHOD to the .jshintrc.
- Added blocks.debug.throwMessage(Message: String, method: (e.g.
  __DEBUG_METHOD, level: [Notice, Warning, Error])
   Level currently only prints an other message and changes the color.
It's not perfect for example I don't like the __DEBUG_METHOD "global" in
the jshintrc but it's the best and cleanest way I could come up with for
now.
Other approches involved scary Estree-Magic and wold fail in case of
errors e.g. in user inserted function (callbacks).
For now it should be enough.