Added check to  make sure that blocks.debug is active, ready and would print something (not paused and not executing).
This doesn't make the performance difference I hoped for because the
optimnization is within a 15-20% mark and that's within the general
js-engine noise. But it's at least fairly consistent and
array.prototype.slice.apply(arguments) is also slow in some enigines.

Optimizing the debug (and normal) build is on my private todo list.