Implemented update for dependency observable arrays.
Overrides normal update action.
Also added a _initialize function to observables that is called to extend
observables in initialization and moved some array specific stuff over to
it.

Also added a _firstExecution method that gets called on the first
execution (get or set) of an observable with the current value as an
argument. Used to extend dependency observables with a ChunkManager
instance.

Also changed the ChunkManager to use obervable._getValue() instead of
observable.__value__.

Closes #119.