Moved test and implemented new property tests.
Added tests for the defaultValue of a blocks.Application.Property to be
the instance of the Model this includes:
  - cloning the defaultValue if if no value is supplied
  - beeing able to pass an object as the value and it gets converted to an
    instance
  - beeing able to pass a instance to the constructor and cloning it.

Also moved a test from spec/mvc/property to spec/mvc/model.
It has nothing to do with properties so I think it's logicaly better
there.

Also added @todos to not implemented tests so I can find them easier
later.