Added "passRaw" option for queries.
passRaw is an object with key = index of the argument and value = boolean.
When the index for an argument is true the execution of that argument is
skiped and the raw argument is passed.
for an example see the if-data-query.

Moved if and if-not data-queries from "hardcoded" queries to "normal"
queries.
Adds 3 lines of extra logic to our data-query handling but I think it's
worth to add a possibility for users write their own (otherwise only
internaly possible) queries.